feat: persist react query cache
