+++ package.json
@@ -16,6 +16,10 @@
     "@expo/vector-icons": "^14.1.0",
     "@react-navigation/native": "^7.1.6",
     "@tanstack/react-query": "^5.85.5",
+    "@tanstack/react-query-persist-client": "^5.0.0",
+    "@tanstack/query-async-storage-persister": "^5.0.0",
+    "@tanstack/query-sync-storage-persister": "^5.0.0",
+    "@react-native-async-storage/async-storage": "^1.23.1",
     "axios": "^1.11.0",
     "expo": "~53.0.22",
     "expo-font": "~13.3.2",
